+++ src/pywnw/__init__.py
@@ -3,6 +3,7 @@
 Modules:
 
 handles -- Provide a class to manage novelWriter handles.
+nw_configuration -- Provide a Configuration class for reading and writing INI files.
 nw_converter -- Provide an novelWriter converter class for yWriter.
 nw_item -- Provide a class for novelWriter items.
 nwx_file -- Provide a class for novelWriter project file representation.
@@ -10,6 +11,7 @@ nwd_file -- Provide a generic class for novelWriter item file representation.
 nwd_novel_file -- Provide a class for novelWriter novel file representation.
 nwd_character_file --  Provide a class for novelWriter character file representation.
 nwd_world_file -- Provide a class for novelWriter world file representation.
+nwd_object_file -- Provide a class for novelWriter object file representation.
 
 Copyright (c) 2022 Peter Triesberger
 For further information see https://github.com/peter88213/PyWriter